feat(dapp-console-api): loginUser endpoint
